requirements.txt
pyodbc==4.0.26
python-decouple==3.1
smart-open==1.8.4
psycopg2==2.8.3
tqdm==4.32.2
